cocaine QOL

+++ modular_skyrat/modules/morenarcotics/items/cocaine.dm
@@ -30,6 +30,8 @@
 	parts = list(/obj/item/reagent_containers/crack = 4)
 	time = 20
 	category = CAT_CHEMISTRY //i might just make a crafting category for drugs at some point
+
+// Should probably give this the edible component at some point
 /obj/item/reagent_containers/cocaine
 	name = "cocaine"
 	desc = "Reenact your favorite scenes from Scarface!"
@@ -39,14 +41,33 @@
 	possible_transfer_amounts = list()
 	list_reagents = list(/datum/reagent/drug/cocaine = 5)
 
-/obj/item/reagent_containers/cocaine/attack(mob/target, mob/user, def_zone)
+/obj/item/reagent_containers/cocaine/proc/snort(mob/living/user)
+	if(!iscarbon(user))
+		return
+	var/covered = ""
+	if(user.is_mouth_covered(head_only = 1))
+		covered = "headgear"
+	else if(user.is_mouth_covered(mask_only = 1))
+		covered = "mask"
+	if(covered)
+		to_chat(user, "<span class='warning'>You have to remove your [covered] first!</span>")
+		return
+	user.visible_message("<span class='notice'[user] starts snorting the [src].</span>")
+	if(do_after(user, 30))
+		to_chat(user, "<span class='notice'>You finish snorting the [src].</span>")
+		if(reagents.total_volume)
+			reagents.trans_to(user, reagents.total_volume, transfered_by = user, methods = INGEST)
+		qdel(src)
+
+/obj/item/reagent_containers/cocaine/attack(mob/target, mob/user)
 	if(target == user)
-		target.visible_message("<span class='notice'>[user] starts snorting the [src].</span>")
-		if(do_after(user,30))
-			to_chat(target, "<span class='notice'>You finish snorting the [src].</span>")
-			if(reagents.total_volume)
-				reagents.trans_to(target, reagents.total_volume, transfered_by = user, methods = INGEST)
-			qdel(src)
+		snort(user)
+
+/obj/item/reagent_containers/cocaine/RightClick(mob/user)
+	. = ..()
+	if(!in_range(user, src) || user.get_active_held_item())
+		return
+	snort(user)
 
 /obj/item/reagent_containers/cocainebrick
 	name = "cocaine brick"
